Fix nasa#2855, addressed unused typedef findings from static analysis.
Most of the changes are replacing native definitions to use the typedef.
Some were removed as they were remnants of previous changes.

+++ modules/evs/config/default_cfe_evs_extern_typedefs.h
@@ -122,7 +122,7 @@ enum CFE_EVS_EventFilter
  *
  * @sa enum CFE_EVS_EventFilter
  */
-typedef uint8 CFE_EVS_EventFilter_Enum_t;
+typedef uint16 CFE_EVS_EventFilter_Enum_t;
 
 /**
  * @brief Label definitions associated with CFE_EVS_EventOutput_Enum_t
